Standart detector

+++ include/detection.hpp
@@ -11,14 +11,14 @@ class Detector {
   static std::shared_ptr<Detector> CreateDetector(const std::string& name);
   virtual bool Init(const std::string& model_file_path) = 0;
   virtual void Detect(const cv::Mat& frame, std::vector<cv::Rect>& objects,
-                      std::vector<double>& scores) = 0;
+                      std::vector<int>& scores) = 0;
 };
 
 class CascadeDetector : public Detector {
 public:
 	bool Init(const std::string& model_file_path);
 	void Detect(const cv::Mat& frame, std::vector<cv::Rect>& objects,
-		std::vector<double>& scores);
+		std::vector<int>& scores);
 
 protected:
 	cv::CascadeClassifier detector;
